Add more consistent use of logging
- Move methods into helper utilites
- Remove debugging log statements
