+++ sbr_ui/routes/authentication.py
@@ -1,7 +1,6 @@
 import uuid
 import logging
 from structlog import wrap_logger
-
 from flask import Blueprint, redirect, url_for, request, flash
 from flask_login import login_required, current_user, logout_user, login_user
